+++ Data/SBV/SMT/SMTLib2.hs
@@ -53,6 +53,7 @@ firstify uniqLen o = prefix o ++ "_" ++ take uniqLen (BC.unpack (B.encode (hash 
   where prefix (SeqOp SBVReverse   {}) = "sbv.reverse"
         prefix (SeqOp SBVZip       {}) = "sbv.zip"
         prefix (SeqOp SBVZipWith   {}) = "sbv.zipWith"
+        prefix (SeqOp SBVReplicate {}) = "sbv.replicate"
         prefix (SeqOp SBVPartition {}) = "sbv.partition"
         prefix (SeqOp SBVMap       {}) = "sbv.map"
         prefix (SeqOp SBVFoldl     {}) = "sbv.foldl"
@@ -343,7 +344,7 @@ cvt ctx curProgInfo kindInfo isSat comments allInputs (_, consts) tbls uis defs 
                  getFuncs (SeqOp (SBVFilter    a     f)) (SeqOp (SBVFilter    x     g)) | [a      ] == [x      ] = Just (f, g)
                  getFuncs (SeqOp (SBVAll       a     f)) (SeqOp (SBVAll       x     g)) | [a      ] == [x      ] = Just (f, g)
                  getFuncs (SeqOp (SBVAny       a     f)) (SeqOp (SBVAny       x     g)) | [a      ] == [x      ] = Just (f, g)
-                 getFuncs _                              _                                                     = Nothing
+                 getFuncs _                              _                                                       = Nothing
 
                  equate :: (Op, (String, SMTLambda), (String, SMTLambda)) -> [String]
                  equate (o, (nm1, f), (nm2, g)) = [ "; Equality for " ++ nm1 ++ " vs. " ++ nm2
@@ -355,14 +356,15 @@ cvt ctx curProgInfo kindInfo isSat comments allInputs (_, consts) tbls uis defs 
                             args   = unwords $ map fst params
                             decls  = unwords ['(' : n ++ " " ++ smtType k ++ ")" | (n, k) <- params]
 
-                            paramsOf (SeqOp (SBVZipWith   k1 k2 _ _)) = [           ("xs", KList k1), ("ys", KList k2)]
-                            paramsOf (SeqOp (SBVPartition k1      _)) = [           ("xs", KList k1)]
-                            paramsOf (SeqOp (SBVMap       k1    _ _)) = [           ("xs", KList k1)]
-                            paramsOf (SeqOp (SBVFoldl     k1 k2   _)) = [("b", k2), ("xs", KList k1)]
-                            paramsOf (SeqOp (SBVFoldr     k1 k2   _)) = [("b", k2), ("xs", KList k1)]
-                            paramsOf (SeqOp (SBVFilter    k1      _)) = [           ("xs", KList k1)]
-                            paramsOf (SeqOp (SBVAll       k1      _)) = [           ("xs", KList k1)]
-                            paramsOf (SeqOp (SBVAny       k1      _)) = [           ("xs", KList k1)]
+                            paramsOf (SeqOp (SBVZipWith   k1 k2 _ _)) = [                   ("xs", KList k1), ("ys", KList k2)]
+                            paramsOf (SeqOp (SBVReplicate k        )) = [("i", KUnbounded), ("e",  k)]
+                            paramsOf (SeqOp (SBVPartition k1      _)) = [                   ("xs", KList k1)]
+                            paramsOf (SeqOp (SBVMap       k1    _ _)) = [                   ("xs", KList k1)]
+                            paramsOf (SeqOp (SBVFoldl     k1 k2   _)) = [("b", k2),         ("xs", KList k1)]
+                            paramsOf (SeqOp (SBVFoldr     k1 k2   _)) = [("b", k2),         ("xs", KList k1)]
+                            paramsOf (SeqOp (SBVFilter    k1      _)) = [                   ("xs", KList k1)]
+                            paramsOf (SeqOp (SBVAll       k1      _)) = [                   ("xs", KList k1)]
+                            paramsOf (SeqOp (SBVAny       k1      _)) = [                   ("xs", KList k1)]
                             paramsOf op                               = error $ "Data.SBV.firstifiedEqualities: Unexpected op: " ++ show op
 
         userDefs = declUserFuns defs
@@ -431,6 +433,7 @@ declSBVFunc cfg op = (nm, comment ++ body)
                  SeqOp (SBVReverse   (KList k))              -> mkSeqRev    (KList k)
                  SeqOp (SBVZip       k1 k2)                  -> mkZip       k1 k2 Nothing
                  SeqOp (SBVZipWith   k1 k2 k3 (SMTLambda f)) -> mkZip       k1 k2 (Just (k3, f))
+                 SeqOp (SBVReplicate k)                      -> mkReplicate k
                  SeqOp (SBVPartition ek       (SMTLambda f)) -> mkPartition ek f
                  SeqOp (SBVMap       k1 k2    (SMTLambda f)) -> mkMap       k1 k2 f
                  SeqOp (SBVFoldl     k1 k2    (SMTLambda f)) -> mkFoldl     k1 k2 f
@@ -456,6 +459,7 @@ declSBVFunc cfg op = (nm, comment ++ body)
                  SeqOp (SBVReverse   k@KList{}) -> shf "reverse"   [k] k
                  SeqOp (SBVZip       a b)       -> shf "zip"       [KList a, KList b] (KList (KTuple [a, b]))
                  SeqOp (SBVZipWith   a b c _)   -> shh "zipWith"   ([a, b], c)  ([KList a, KList b], KList c)
+                 SeqOp (SBVReplicate a)         -> shf "replicate" [KUnbounded, a] (KList a)
                  SeqOp (SBVPartition a     _)   -> shh "partition" ([a], KBool) ([KList a], KTuple [KList a, KList a])
                  SeqOp (SBVMap       a b   _)   -> shh "map"       ([a], b)     ([KList a], KList b)
                  SeqOp (SBVFoldl     a b   _)   -> shh "foldl"     ([b, a], b)  ([b, KList a], b)
@@ -591,6 +595,15 @@ declSBVFunc cfg op = (nm, comment ++ body)
                      ]
          where tla  = smtType (KList a)
                tlla = smtType (KList (KList a))
+
+        -- Int -> a -> [a]
+        mkReplicate a = [ "(define-fun-rec " ++ nm ++ " ((i Int) (e " ++ ta ++ ")) " ++ tla
+                        , "                (ite (<= i 0)"
+                        , "                     " ++ empty tla
+                        , "                     (seq.++ (seq.unit e) (" ++ nm ++ " (- i 1) e))))"
+                        ]
+         where ta  = smtType a
+               tla = smtType (KList a)
 
 -- | Declare new sorts
 declSort :: (String, Maybe [String]) -> [String]
@@ -1261,21 +1274,22 @@ cvtExp cfg curProgInfo caps rm tableMap expr@(SBVApp _ arguments) = sh expr
         sh (SBVApp (RegExOp o@RegExEq{})  []) = show o
         sh (SBVApp (RegExOp o@RegExNEq{}) []) = show o
 
-        -- Reverse and higher order functions are special
-        sh (SBVApp o@(SeqOp SBVReverse{})   args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
-        sh (SBVApp o@(SeqOp SBVZip{})       args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
-        sh (SBVApp o@(SeqOp SBVZipWith{})   args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
-        sh (SBVApp o@(SeqOp SBVPartition{}) args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
-        sh (SBVApp o@(SeqOp SBVReverse{})   args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
-        sh (SBVApp o@(SeqOp SBVMap{})       args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
-        sh (SBVApp o@(SeqOp SBVFoldl{})     args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
-        sh (SBVApp o@(SeqOp SBVFoldr{})     args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
-        sh (SBVApp o@(SeqOp SBVFilter{})    args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
-        sh (SBVApp o@(SeqOp SBVAll{} )      args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
-        sh (SBVApp o@(SeqOp SBVAny{} )      args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
-        sh (SBVApp o@(SeqOp SBVConcat{})    args) = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        -- Reverse and higher order functions are special. Reverse is supported for strings, not others.
+        sh (SBVApp o@(SeqOp SBVReverse{})   args)                = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVZip{})       args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVZipWith{})   args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVReplicate{}) args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVPartition{}) args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVReverse{})   args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVMap{})       args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVFoldl{})     args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVFoldr{})     args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVFilter{})    args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVAll{} )      args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVAny{} )      args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp o@(SeqOp SBVConcat{})    args) | not stringOp = "(" ++ firstifiedName o ++ " " ++ unwords (map cvtSV args) ++ ")"
 
-        sh (SBVApp (SeqOp op) args) = "(" ++ show op ++ " " ++ unwords (map cvtSV args) ++ ")"
+        sh (SBVApp (SeqOp op) args) = "(" ++ stringOrSequence stringOp op ++ " " ++ unwords (map cvtSV args) ++ ")"
 
         sh (SBVApp (SetOp SetEqual)      args)   = "(= "      ++ unwords (map cvtSV args) ++ ")"
         sh (SBVApp (SetOp SetMember)     [e, s]) = "(select " ++ cvtSV s ++ " " ++ cvtSV e ++ ")"
@@ -1710,3 +1724,37 @@ setSMTOption cfg = set
         smtBool :: Bool -> String
         smtBool True  = "true"
         smtBool False = "false"
+
+-- Pick the correct op if we have a string
+stringOrSequence :: Bool -> SeqOp -> String
+stringOrSequence False op = show op
+stringOrSequence True  op = case op of
+                              SeqConcat       -> show StrConcat
+                              SeqLen          -> show StrLen
+                              SeqUnit         -> show StrUnit
+                              SeqNth          -> show StrNth
+                              SeqSubseq       -> show StrSubstr
+                              SeqIndexOf      -> show StrIndexOf
+                              SeqContains     -> show StrContains
+                              SeqPrefixOf     -> show StrPrefixOf
+                              SeqSuffixOf     -> show StrSuffixOf
+                              SeqReplace      -> show StrReplace
+                              SBVReverse   {} -> tbd
+                              SBVZip       {} -> tbd
+                              SBVReplicate {} -> tbd
+                              SBVZipWith   {} -> tbd
+                              SBVPartition {} -> tbd
+                              SBVMap       {} -> tbd
+                              SBVFoldl     {} -> tbd
+                              SBVFoldr     {} -> tbd
+                              SBVFilter    {} -> tbd
+                              SBVAll       {} -> tbd
+                              SBVAny       {} -> tbd
+                              SBVConcat    {} -> tbd
+  where tbd = error $ unlines [ "***"
+                              , "*** Use of sequence operator " ++ show op ++ " on strings isn't supported yet."
+                              , "***"
+                              , "*** Please cast this in terms of operators in Data.SBV.String. If a polymorphic"
+                              , "*** use case is needed that works uniformly on both strings and arbitrary"
+                              , "*** sequences, please report this as a feature request."
+                              ]