+++ Documentation/SBV/Examples/KnuckleDragger/MergeSort.hs
@@ -9,6 +9,7 @@
 -- Proving merge sort correct.
 -----------------------------------------------------------------------------
 
+{-# LANGUAGE CPP                 #-}
 {-# LANGUAGE DataKinds           #-}
 {-# LANGUAGE TypeAbstractions    #-}
 {-# LANGUAGE TypeApplications    #-}
@@ -18,16 +19,25 @@
 
 module Documentation.SBV.Examples.KnuckleDragger.MergeSort where
 
+import Prelude hiding (null, length, head, tail, elem, splitAt, (++), take, drop)
+import Data.Proxy
+
 import Data.SBV
+import Data.SBV.List
 import Data.SBV.Tools.KnuckleDragger
 
-import Prelude hiding (null, length, head, tail, elem, splitAt, (++), take, drop)
-import Data.SBV.List
+import qualified Documentation.SBV.Examples.KnuckleDragger.SortHelpers as SH
+
+#ifdef DOCTEST
+-- $setup
+-- >>> :set -XTypeApplications
+-- >>> import Data.Proxy
+#endif
 
 -- * Merge sort
 
 -- | Merge two already sorted lists into another
-merge :: SList Integer -> SList Integer -> SList Integer
+merge :: (Ord a, SymVal a) => SList a -> SList a -> SList a
 merge = smtFunction "merge" $ \l r -> ite (null l) r
                                     $ ite (null r) l
                                     $ let (a, as) = uncons l
@@ -35,30 +45,10 @@ merge = smtFunction "merge" $ \l r -> ite (null l) r
                                       in ite (a .<= b) (a .: merge as r) (b .: merge l bs)
 
 -- | Merge sort, using 'merge' above to successively sort halved input
-mergeSort :: SList Integer -> SList Integer
+mergeSort :: (Ord a, SymVal a) => SList a -> SList a
 mergeSort = smtFunction "mergeSort" $ \l -> ite (length l .<= 1) l
                                               $ let (h1, h2) = splitAt (length l `sEDiv` 2) l
                                                 in merge (mergeSort h1) (mergeSort h2)
--- * Helper functions
-
--- | A predicate testing whether a given list is non-decreasing.
-nonDecreasing :: SList Integer -> SBool
-nonDecreasing = smtFunction "nonDecreasing" $ \l ->  null l .|| null (tail l)
-                                                 .|| let (x, l') = uncons l
-                                                         (y, _)  = uncons l'
-                                                     in x .<= y .&& nonDecreasing l'
-
--- | Count the number of occurrences of an element in a list
-count :: SInteger -> SList Integer -> SInteger
-count = smtFunction "count" $ \e l -> ite (null l)
-                                          0
-                                          (let (x, xs) = uncons l
-                                               cxs     = count e xs
-                                           in ite (e .== x) (1 + cxs) cxs)
-
--- | Are two lists permutations of each other?
-isPermutation :: SList Integer -> SList Integer -> SBool
-isPermutation xs ys = quantifiedBool (\(Forall @"x" x) -> count x xs .== count x ys)
 
 -- * Correctness proof
 
@@ -66,8 +56,20 @@ isPermutation xs ys = quantifiedBool (\(Forall @"x" x) -> count x xs .== count x
 --
 -- We have:
 --
--- >>> correctness
--- Lemma: nonDecrInsert                                        Q.E.D.
+-- >>> correctness (Proxy @Integer)
+-- Lemma: nonDecrInsert                    Q.E.D.
+-- Inductive lemma: countAppend @Integer
+--   Step: Base                            Q.E.D.
+--   Step: 1                               Q.E.D.
+--   Step: 2 (unfold count)                Q.E.D.
+--   Step: 3                               Q.E.D.
+--   Step: 4 (simplify)                    Q.E.D.
+--   Result:                               Q.E.D.
+-- Lemma: take_drop @Integer               Q.E.D.
+-- Lemma: takeDropCount @Integer
+--   Step: 1                               Q.E.D.
+--   Step: 2                               Q.E.D.
+--   Result:                               Q.E.D.
 -- Inductive lemma (strong): mergeKeepsSort
 --   Step: Measure is non-negative                             Q.E.D.
 --   Step: 1 (4 way full case split)
@@ -105,18 +107,6 @@ isPermutation xs ys = quantifiedBool (\(Forall @"x" x) -> count x xs .== count x
 --     Step: 1.4.6 (unfold count in reverse, twice)            Q.E.D.
 --     Step: 1.4.7 (simplify)                                  Q.E.D.
 --   Result:                                                   Q.E.D.
--- Inductive lemma: countAppend
---   Step: Base                                                Q.E.D.
---   Step: 1                                                   Q.E.D.
---   Step: 2 (unfold count)                                    Q.E.D.
---   Step: 3                                                   Q.E.D.
---   Step: 4 (simplify)                                        Q.E.D.
---   Result:                                                   Q.E.D.
--- Lemma: take_drop                                            Q.E.D.
--- Lemma: takeDropCount
---   Step: 1                                                   Q.E.D.
---   Step: 2                                                   Q.E.D.
---   Result:                                                   Q.E.D.
 -- Inductive lemma (strong): sortIsPermutation
 --   Step: Measure is non-negative                             Q.E.D.
 --   Step: 1 (2 way full case split)
@@ -130,22 +120,27 @@ isPermutation xs ys = quantifiedBool (\(Forall @"x" x) -> count x xs .== count x
 --   Result:                                                   Q.E.D.
 -- Lemma: mergeSortIsCorrect                                   Q.E.D.
 -- [Proven] mergeSortIsCorrect
-correctness :: IO Proof
-correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
+correctness :: forall a. (Ord a, SymVal a) => Proxy a -> IO Proof
+correctness _ = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
 
     --------------------------------------------------------------------------------------------
-    -- Part I. Prove that the output of merge sort is non-decreasing.
+    -- Part I. Import helper lemmas, definitions
     --------------------------------------------------------------------------------------------
+    let nonDecreasing = SH.nonDecreasing @a
+        isPermutation = SH.isPermutation @a
+        count         = SH.count         @a
 
-    nonDecrIns <- lemma "nonDecrInsert"
-                        (\(Forall @"x" x) (Forall @"ys" ys) -> nonDecreasing ys .&& sNot (null ys) .&& x .<= head ys
-                                                           .=> nonDecreasing (x .: ys))
-                        []
+    nonDecrIns    <- use $ SH.nonDecrIns    (Proxy @a)
+    takeDropCount <- use $ SH.takeDropCount (Proxy @a)
+
+    --------------------------------------------------------------------------------------------
+    -- Part II. Prove that the output of merge sort is non-decreasing.
+    --------------------------------------------------------------------------------------------
 
     mergeKeepsSort <-
         sInductWith cvc5 "mergeKeepsSort"
            (\(Forall @"xs" xs) (Forall @"ys" ys) -> nonDecreasing xs .&& nonDecreasing ys .=> nonDecreasing (merge xs ys))
-           (\(xs :: SList Integer) (ys :: SList Integer) -> (length xs, length ys)) $
+           (\xs ys -> (length @a xs, length @a ys)) $
            \ih xs ys -> [nonDecreasing xs, nonDecreasing ys]
                      |- split2 (xs, ys)
                                trivial           -- when both xs and ys are empty.  Trivial.
@@ -175,7 +170,7 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
     sortNonDecreasing <-
         sInduct "sortNonDecreasing"
                 (\(Forall @"xs" xs) -> nonDecreasing (mergeSort xs))
-                (length @Integer) $
+                (length @a) $
                 \ih xs -> [] |- split xs
                                       qed
                                       (\e es -> nonDecreasing (mergeSort (e .: es))
@@ -200,12 +195,13 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
                                              =: qed)
 
     --------------------------------------------------------------------------------------------
-    -- Part II. Prove that the output of merge sort is a permuation of its input
+    -- Part III. Prove that the output of merge sort is a permuation of its input
     --------------------------------------------------------------------------------------------
     mergeCount <-
         sInduct "mergeCount"
-                (\(Forall @"xs" xs) (Forall @"ys" ys) (Forall @"e" e) -> count e (merge xs ys) .== count e xs + count e ys)
-                (\(xs :: SList Integer) (ys :: SList Integer) (_e :: SInteger) -> (length xs, length ys)) $
+                (\(Forall @"xs" xs) (Forall @"ys" ys) (Forall @"e" (e :: SBV a)) ->
+                        count e (merge xs ys) .== count e xs + count e ys)
+                (\xs ys (_e :: SBV a) -> (length @a xs, length @a ys)) $
                 \ih as bs e -> [] |-
                         split2 (as, bs)
                                trivial
@@ -240,38 +236,10 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
                                                  =: count e (x .: xs) + count e (y .: ys)
                                                  =: qed)
 
-    countAppend <-
-      induct "countAppend"
-             (\(Forall @"xs" xs) (Forall @"ys" ys) (Forall @"e" e) -> count e (xs ++ ys) .== count e xs + count e ys) $
-             \ih x xs ys e -> [] |- count e ((x .: xs) ++ ys)
-                                 =: count e (x .: (xs ++ ys))
-                                 ?? "unfold count"
-                                 =: (let r = count e (xs ++ ys) in ite (e .== x) (1+r) r)
-                                 ?? ih `at` (Inst @"ys" ys, Inst @"e" e)
-                                 =: (let r = count e xs + count e ys in ite (e .== x) (1+r) r)
-                                 ?? "simplify"
-                                 =: count e (x .: xs) + count e ys
-                                 =: qed
-
-    takeDropCount <- do
-
-       takeDrop <- lemma "take_drop"
-                         (\(Forall @"n" n) (Forall @"xs" (xs :: SList Integer)) -> take n xs ++ drop n xs .== xs)
-                         []
-
-       calc "takeDropCount"
-            (\(Forall @"xs" xs) (Forall @"n" n) (Forall @"e" e) -> count e (take n xs) + count e (drop n xs) .== count e xs) $
-            \xs n e -> [] |- count e (take n xs) + count e (drop n xs)
-                          ?? countAppend `at` (Inst @"xs" (take n xs), Inst @"ys" (drop n xs), Inst @"e" e)
-                          =: count e (take n xs ++ drop n xs)
-                          ?? takeDrop
-                          =: count e xs
-                          =: qed
-
     sortIsPermutation <-
         sInductWith cvc5 "sortIsPermutation"
-                (\(Forall @"xs" xs) (Forall @"e" e) -> count e xs .== count e (mergeSort xs))
-                (\(xs :: SList Integer) (_e :: SInteger) -> length xs) $
+                (\(Forall @"xs" (xs :: SList a)) (Forall @"e" e) -> count e xs .== count e (mergeSort xs))
+                (\xs (_e :: SBV a) -> length @a xs) $
                 \ih as e -> [] |- split as
                                         qed
                                         (\x xs -> count e (mergeSort (x .: xs))