+++ Documentation/SBV/Examples/KnuckleDragger/QuickSort.hs
@@ -12,6 +12,7 @@
 -- back in 1989.
 -----------------------------------------------------------------------------
 
+{-# LANGUAGE CPP                 #-}
 {-# LANGUAGE DataKinds           #-}
 {-# LANGUAGE TypeAbstractions    #-}
 {-# LANGUAGE TypeApplications    #-}
@@ -23,16 +24,25 @@ module Documentation.SBV.Examples.KnuckleDragger.QuickSort where
 
 import Prelude hiding (null, length, (++), tail, all, fst, snd, elem)
 import Control.Monad.Trans (liftIO)
+import Data.Proxy
 
 import Data.SBV
 import Data.SBV.List hiding (partition)
 import Data.SBV.Tuple
 import Data.SBV.Tools.KnuckleDragger
 
+import qualified Documentation.SBV.Examples.KnuckleDragger.SortHelpers as SH
+
+#ifdef DOCTEST
+-- $setup
+-- >>> :set -XTypeApplications
+-- >>> import Data.Proxy
+#endif
+
 -- * Quick sort
 
 -- | Quick-sort, using the first element as pivot.
-quickSort :: SList Integer -> SList Integer
+quickSort :: (Ord a, SymVal a) => SList a -> SList a
 quickSort = smtFunction "quickSort" $ \l -> ite (null l)
                                                 nil
                                                 (let (x,  xs) = uncons l
@@ -42,7 +52,7 @@ quickSort = smtFunction "quickSort" $ \l -> ite (null l)
 -- | We define @partition@ as an explicit function. Unfortunately, we can't just replace this
 -- with @\pivot xs -> Data.List.SBV.partition (.< pivot) xs@ because that would create a firstified version of partition
 -- with a free-variable captured, which isn't supported due to higher-order limitations in SMTLib.
-partition :: SInteger -> SList Integer -> STuple [Integer] [Integer]
+partition :: (Ord a, SymVal a) => SBV a -> SList a -> STuple [a] [a]
 partition = smtFunction "partition" $ \pivot xs -> ite (null xs)
                                                        (tuple (nil, nil))
                                                        (let (a,  as) = uncons xs
@@ -51,34 +61,54 @@ partition = smtFunction "partition" $ \pivot xs -> ite (null xs)
                                                                (tuple (a .: lo, hi))
                                                                (tuple (lo, a .: hi)))
 
--- * Helper functions
-
--- | A predicate testing whether a given list is non-decreasing.
-nonDecreasing :: SList Integer -> SBool
-nonDecreasing = smtFunction "nonDecreasing" $ \l ->  null l .|| null (tail l)
-                                                 .|| let (x, l') = uncons l
-                                                         (y, _)  = uncons l'
-                                                     in x .<= y .&& nonDecreasing l'
-
--- | Count the number of occurrences of an element in a list
-count :: SInteger -> SList Integer -> SInteger
-count = smtFunction "count" $ \e l -> ite (null l)
-                                          0
-                                          (let (x, xs) = uncons l
-                                               cxs     = count e xs
-                                           in ite (e .== x) (1 + cxs) cxs)
-
--- | Are two lists permutations of each other?
-isPermutation :: SList Integer -> SList Integer -> SBool
-isPermutation xs ys = quantifiedBool (\(Forall @"x" x) -> count x xs .== count x ys)
-
 -- * Correctness proof
 
 -- | Correctness of quick-sort.
 --
 -- We have:
 --
--- >>> correctness
+-- >>> correctness (Proxy @Integer)
+-- Inductive lemma: countAppend @Integer
+--   Step: Base                            Q.E.D.
+--   Step: 1                               Q.E.D.
+--   Step: 2 (unfold count)                Q.E.D.
+--   Step: 3                               Q.E.D.
+--   Step: 4 (simplify)                    Q.E.D.
+--   Result:                               Q.E.D.
+-- Inductive lemma: countNonNegative @Integer
+--   Step: Base                            Q.E.D.
+--   Step: 1 (2 way case split)
+--     Step: 1.1.1                         Q.E.D.
+--     Step: 1.1.2                         Q.E.D.
+--     Step: 1.2.1                         Q.E.D.
+--     Step: 1.2.2                         Q.E.D.
+--     Step: 1.Completeness                Q.E.D.
+--   Result:                               Q.E.D.
+-- Inductive lemma: countElem @Integer
+--   Step: Base                            Q.E.D.
+--   Step: 1 (2 way case split)
+--     Step: 1.1.1                         Q.E.D.
+--     Step: 1.1.2                         Q.E.D.
+--     Step: 1.2.1                         Q.E.D.
+--     Step: 1.2.2                         Q.E.D.
+--     Step: 1.Completeness                Q.E.D.
+--   Result:                               Q.E.D.
+-- Inductive lemma: elemCount @Integer
+--   Step: Base                            Q.E.D.
+--   Step: 1 (2 way case split)
+--     Step: 1.1                           Q.E.D.
+--     Step: 1.2.1                         Q.E.D.
+--     Step: 1.2.2                         Q.E.D.
+--     Step: 1.Completeness                Q.E.D.
+--   Result:                               Q.E.D.
+-- Lemma: sublistCorrect @Integer
+--   Step: 1                               Q.E.D.
+--   Result:                               Q.E.D.
+-- Lemma: sublistElem @Integer
+--   Step: 1                               Q.E.D.
+--   Result:                               Q.E.D.
+-- Lemma: sublistTail @Integer             Q.E.D.
+-- Lemma: permutationImpliesSublist @IntegerQ.E.D.
 -- Inductive lemma: lltCorrect
 --   Step: Base                                                Q.E.D.
 --   Step: 1                                                   Q.E.D.
@@ -87,40 +117,6 @@ isPermutation xs ys = quantifiedBool (\(Forall @"x" x) -> count x xs .== count x
 --   Step: Base                                                Q.E.D.
 --   Step: 1                                                   Q.E.D.
 --   Result:                                                   Q.E.D.
--- Inductive lemma: countNonNegative
---   Step: Base                                                Q.E.D.
---   Step: 1 (2 way case split)
---     Step: 1.1.1                                             Q.E.D.
---     Step: 1.1.2                                             Q.E.D.
---     Step: 1.2.1                                             Q.E.D.
---     Step: 1.2.2                                             Q.E.D.
---     Step: 1.Completeness                                    Q.E.D.
---   Result:                                                   Q.E.D.
--- Inductive lemma: countElem
---   Step: Base                                                Q.E.D.
---   Step: 1 (2 way case split)
---     Step: 1.1.1                                             Q.E.D.
---     Step: 1.1.2                                             Q.E.D.
---     Step: 1.2.1                                             Q.E.D.
---     Step: 1.2.2                                             Q.E.D.
---     Step: 1.Completeness                                    Q.E.D.
---   Result:                                                   Q.E.D.
--- Inductive lemma: elemCount
---   Step: Base                                                Q.E.D.
---   Step: 1 (2 way case split)
---     Step: 1.1                                               Q.E.D.
---     Step: 1.2.1                                             Q.E.D.
---     Step: 1.2.2                                             Q.E.D.
---     Step: 1.Completeness                                    Q.E.D.
---   Result:                                                   Q.E.D.
--- Lemma: sublistCorrect
---   Step: 1                                                   Q.E.D.
---   Result:                                                   Q.E.D.
--- Lemma: sublistElem
---   Step: 1                                                   Q.E.D.
---   Result:                                                   Q.E.D.
--- Lemma: sublistTail                                          Q.E.D.
--- Lemma: permutationImpliesSublist                            Q.E.D.
 -- Inductive lemma: lltSublist
 --   Step: Base                                                Q.E.D.
 --   Step: 1                                                   Q.E.D.
@@ -164,13 +160,6 @@ isPermutation xs ys = quantifiedBool (\(Forall @"x" x) -> count x xs .== count x
 --     Step: 1.2.1                                             Q.E.D.
 --     Step: 1.2.2 (simplify)                                  Q.E.D.
 --     Step: 1.2.3                                             Q.E.D.
---   Result:                                                   Q.E.D.
--- Inductive lemma: countAppend
---   Step: Base                                                Q.E.D.
---   Step: 1                                                   Q.E.D.
---   Step: 2 (unfold count)                                    Q.E.D.
---   Step: 3                                                   Q.E.D.
---   Step: 4 (simplify)                                        Q.E.D.
 --   Result:                                                   Q.E.D.
 -- Inductive lemma: countPartition
 --   Step: Base                                                Q.E.D.
@@ -216,7 +205,7 @@ isPermutation xs ys = quantifiedBool (\(Forall @"x" x) -> count x xs .== count x
 --     Step: 1.2.3 (push nonDecreasing down)                   Q.E.D.
 --     Step: 1.2.4                                             Q.E.D.
 --   Result:                                                   Q.E.D.
--- Lemma: quickSortIsCorrect                                   Q.E.D.
+-- Lemma: quickSortIsCorrect @Integer                          Q.E.D.
 -- == Proof tree:
 -- quickSortIsCorrect
 --  ├╴sortIsPermutation
@@ -233,37 +222,39 @@ isPermutation xs ys = quantifiedBool (\(Forall @"x" x) -> count x xs .== count x
 --     ├╴sortIsPermutation (x2)
 --     ├╴lltPermutation
 --     │  ├╴lltSublist
---     │  │  ├╴sublistElem
---     │  │  │  └╴sublistCorrect
---     │  │  │     ├╴countElem
---     │  │  │     │  └╴countNonNegative
---     │  │  │     └╴elemCount
---     │  │  ├╴lltCorrect
---     │  │  └╴sublistTail
+--     │  │  └╴sublistElem (x3)
 --     │  └╴permutationImpliesSublist
 --     ├╴lgePermutation
 --     │  ├╴lgeSublist
---     │  │  ├╴sublistElem
---     │  │  ├╴lgeCorrect
---     │  │  └╴sublistTail
+--     │  │  ├╴sublistElem (x2)
+--     │  │  └╴lgeCorrect
 --     │  └╴permutationImpliesSublist
 --     └╴nonDecreasingMerge
--- [Proven] quickSortIsCorrect
-correctness :: IO Proof
-correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
+-- [Proven] quickSortIsCorrect @Integer
+correctness :: forall a. (Ord a, SymVal a) => Proxy a -> IO Proof
+correctness p = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
+
+  --------------------------------------------------------------------------------------------
+  -- Part I. Import helper lemmas, definitions
+  --------------------------------------------------------------------------------------------
+  let count         = SH.count         @a
+      isPermutation = SH.isPermutation @a
+      nonDecreasing = SH.nonDecreasing @a
+      sublist       = SH.sublist       @a
+
+  countAppend               <- use $ SH.countAppend               p
+  sublistElem               <- use $ SH.sublistElem               p
+  sublistTail               <- use $ SH.sublistTail               p
+  permutationImpliesSublist <- use $ SH.permutationImpliesSublist p
 
   ---------------------------------------------------------------------------------------------------
-  -- Part I. Formalizing less-than/greater-than-or-equal over lists and relationship to permutations
+  -- Part II. Formalizing less-than/greater-than-or-equal over lists and relationship to permutations
   ---------------------------------------------------------------------------------------------------
   -- llt: list less-than:     all the elements are <  pivot
   -- lge: list greater-equal: all the elements are >= pivot
-  let llt, lge :: SInteger -> SList Integer -> SBool
+  let llt, lge :: SBV a -> SList a -> SBool
       llt = smtFunction "llt" $ \pivot l -> null l .|| let (x, xs) = uncons l in x .<  pivot .&& llt pivot xs
       lge = smtFunction "lge" $ \pivot l -> null l .|| let (x, xs) = uncons l in x .>= pivot .&& lge pivot xs
-
-      -- Sublist relationship
-      sublist :: SList Integer -> SList Integer -> SBool
-      sublist xs ys = quantifiedBool (\(Forall @"e" e) -> count e xs .> 0 .=> count e ys .> 0)
 
   -- llt correctness
   lltCorrect <-
@@ -284,77 +275,6 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
                              ?? ih
                              =: sTrue
                              =: qed
-
-  -- count is always non-negative
-  countNonNegative <- induct "countNonNegative"
-                             (\(Forall @"xs" xs) (Forall @"e" e) -> count e xs .>= 0) $
-                             \ih x xs e -> [] |- count e (x .: xs) .>= 0
-                                              =: cases [ e .== x ==> 1 + count e xs .>= 0
-                                                                  ?? ih
-                                                                  =: sTrue
-                                                                  =: qed
-                                                       , e ./= x ==> count e xs .>= 0
-                                                                  ?? ih
-                                                                  =: sTrue
-                                                                  =: qed
-                                                       ]
-
-  -- relationship between count and elem, forward direction
-  countElem <- induct "countElem"
-                      (\(Forall @"xs" xs) (Forall @"e" e) -> e `elem` xs .=> count e xs .> 0) $
-                      \ih x xs e -> [e `elem` (x .: xs)]
-                                 |- count e (x .: xs) .> 0
-                                 =: cases [ e .== x ==> 1 + count e xs .> 0
-                                                     ?? countNonNegative
-                                                     =: sTrue
-                                                     =: qed
-                                          , e ./= x ==> count e xs .> 0
-                                                     ?? ih
-                                                     =: sTrue
-                                                     =: qed
-                                          ]
-
-  -- relationship between count and elem, backwards direction
-  elemCount <- induct "elemCount"
-                      (\(Forall @"xs" xs) (Forall @"e" e) -> count e xs .> 0 .=> e `elem` xs) $
-                      \ih x xs e -> [count e xs .> 0]
-                                 |- e `elem` (x .: xs)
-                                 =: cases [ e .== x ==> trivial
-                                          , e ./= x ==> e `elem` xs
-                                                     ?? ih
-                                                     =: sTrue
-                                                     =: qed
-                                          ]
-
-  -- sublist correctness
-  sublistCorrect <- calc "sublistCorrect"
-                          (\(Forall @"xs" xs) (Forall @"ys" ys) (Forall @"x" x) -> xs `sublist` ys .&& x `elem` xs .=> x `elem` ys) $
-                          \xs ys x -> [xs `sublist` ys, x `elem` xs]
-                                   |- x `elem` ys
-                                   ?? [ countElem `at` (Inst @"xs" xs, Inst @"e" x)
-                                      , elemCount `at` (Inst @"xs" ys, Inst @"e" x)
-                                      ]
-                                   =: sTrue
-                                   =: qed
-
-  -- If one list is a sublist of another, then its head is an elem
-  sublistElem <- calc "sublistElem"
-                       (\(Forall @"x" x) (Forall @"xs" xs) (Forall @"ys" ys) -> (x .: xs) `sublist` ys .=> x `elem` ys) $
-                       \x xs ys -> [(x .: xs) `sublist` ys]
-                                |- x `elem` ys
-                                ?? sublistCorrect `at` (Inst @"xs" (x .: xs), Inst @"ys" ys, Inst @"x" x)
-                                =: sTrue
-                                =: qed
-
-  -- If one list is a sublist of another so is its tail
-  sublistTail <- lemma "sublistTail"
-                       (\(Forall @"x" x) (Forall @"xs" xs) (Forall @"ys" ys) -> (x .: xs) `sublist` ys .=> xs `sublist` ys)
-                       []
-
-  -- Permutation implies sublist
-  permutationImpliesSublist <- lemma "permutationImpliesSublist"
-                                    (\(Forall @"xs" xs) (Forall @"ys" ys) -> isPermutation xs ys .=> xs `sublist` ys)
-                                    []
 
   -- If a value is less than all the elements in a list, then it is also less than all the elements of any sublist of it
   lltSublist <-
@@ -421,7 +341,7 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
                         =: qed
 
   --------------------------------------------------------------------------------------------
-  -- Part II. Helper lemmas for partition
+  -- Part III. Helper lemmas for partition
   --------------------------------------------------------------------------------------------
 
   -- The first element of the partition produces all smaller elements
@@ -456,9 +376,9 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
 
   -- The first element of partition does not increase in size
   partitionNotLongerFst <- sInduct "partitionNotLongerFst"
-     (\(Forall @"l" l) (Forall @"pivot" pivot) -> length (fst (partition pivot l)) .<= length l)
-     (\l (_ :: SInteger) -> length @Integer l) $
-     \ih l pivot -> [] |- length (fst (partition pivot l)) .<= length l
+     (\(Forall @"l" l) (Forall @"pivot" pivot) -> length (fst (partition @a pivot l)) .<= length l)
+     (\l (_ :: SBV a) -> length @a l) $
+     \ih l pivot -> [] |- length (fst (partition @a pivot l)) .<= length l
                        =: split l trivial
                                 (\a as -> let lo = fst (partition pivot as)
                                        in ite (a .< pivot)
@@ -474,9 +394,9 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
 
   -- The second element of partition does not increase in size
   partitionNotLongerSnd <- sInduct "partitionNotLongerSnd"
-     (\(Forall @"l" l) (Forall @"pivot" pivot) -> length (snd (partition pivot l)) .<= length l)
-     (\l (_ :: SInteger) -> length @Integer l) $
-     \ih l pivot -> [] |- length (snd (partition pivot l)) .<= length l
+     (\(Forall @"l" l) (Forall @"pivot" pivot) -> length (snd (partition @a pivot l)) .<= length l)
+     (\l (_ :: SBV a) -> length @a l) $
+     \ih l pivot -> [] |- length (snd (partition @a pivot l)) .<= length l
                        =: split l trivial
                                 (\a as -> let hi = snd (partition pivot as)
                                        in ite (a .< pivot)
@@ -491,25 +411,11 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
                                        =: qed)
 
   --------------------------------------------------------------------------------------------
-  -- Part III. Helper lemmas for count
+  -- Part IV. Helper lemmas for count
   --------------------------------------------------------------------------------------------
 
-  -- Count distributes over append
-  countAppend <-
-      induct "countAppend"
-             (\(Forall @"xs" xs) (Forall @"ys" ys) (Forall @"e" e) -> count e (xs ++ ys) .== count e xs + count e ys) $
-             \ih x xs ys e -> [] |- count e ((x .: xs) ++ ys)
-                                 =: count e (x .: (xs ++ ys))
-                                 ?? "unfold count"
-                                 =: (let r = count e (xs ++ ys) in ite (e .== x) (1+r) r)
-                                 ?? ih `at` (Inst @"ys" ys, Inst @"e" e)
-                                 =: (let r = count e xs + count e ys in ite (e .== x) (1+r) r)
-                                 ?? "simplify"
-                                 =: count e (x .: xs) + count e ys
-                                 =: qed
-
   -- Count is preserved over partition
-  let countTuple :: SInteger -> STuple [Integer] [Integer] -> SInteger
+  let countTuple :: SBV a -> STuple [a] [a] -> SInteger
       countTuple e xsys = count e xs + count e ys
         where (xs, ys) = untuple xsys
 
@@ -546,13 +452,13 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
                                        =: qed
                             ]
   --------------------------------------------------------------------------------------------
-  -- Part IV. Prove that the output of quick sort is a permutation of its input
+  -- Part V. Prove that the output of quick sort is a permutation of its input
   --------------------------------------------------------------------------------------------
 
   sortCountsMatch <-
      sInduct "sortCountsMatch"
-             (\(Forall @"xs" xs) (Forall @"e" e) -> count e xs .== count e (quickSort xs))
-             (\xs (_ :: SInteger) -> length @Integer xs) $
+             (\(Forall @"xs" xs) (Forall @"e" (e :: SBV a)) -> count e xs .== count e (quickSort xs))
+             (\xs (_ :: SBV a) -> length @a xs) $
              \ih xs e ->
                 [] |- count e (quickSort xs)
                    =: split xs trivial
@@ -567,17 +473,15 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
                                    =: count e (quickSort lo) + count e (singleton a ++ quickSort hi)
                                    ?? countAppend `at` (Inst @"xs" (singleton a), Inst @"ys" (quickSort hi), Inst @"e" e)
                                    =: count e (quickSort lo) + count e (singleton a) + count e (quickSort hi)
-                                   ?? [ hprf  $ ih                    `at` (Inst @"xs" lo, Inst @"e" e)
-                                      , hprf  $ partitionNotLongerFst `at` (Inst @"l"  as, Inst @"pivot" a)
-                                      , hasm  $ xs .== a .: as
-                                      , hcmnt "IH on lo"
+                                   ?? [ ih                    `at` (Inst @"xs" lo, Inst @"e" e)
+                                      , partitionNotLongerFst `at` (Inst @"l"  as, Inst @"pivot" a)
                                       ]
+                                   ?? "IH on lo"
                                    =: count e lo + count e (singleton a) + count e (quickSort hi)
-                                   ?? [ hprf  $ ih                    `at` (Inst @"xs" hi, Inst @"e" e)
-                                      , hprf  $ partitionNotLongerSnd `at` (Inst @"l"  as, Inst @"pivot" a)
-                                      , hasm  $ xs .== a .: as
-                                      , hcmnt "IH on hi"
+                                   ?? [ ih                    `at` (Inst @"xs" hi, Inst @"e" e)
+                                      , partitionNotLongerSnd `at` (Inst @"l"  as, Inst @"pivot" a)
                                       ]
+                                   ?? "IH on hi"
                                    =: count e lo + count e (singleton a) + count e hi
                                    ?? countPartition `at` (Inst @"xs" as, Inst @"pivot" a, Inst @"e" e)
                                    =: count e xs
@@ -586,7 +490,7 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
   sortIsPermutation <- lemma "sortIsPermutation" (\(Forall @"xs" xs) -> isPermutation xs (quickSort xs)) [sortCountsMatch]
 
   --------------------------------------------------------------------------------------------
-  -- Part V. Helper lemmas for nonDecreasing
+  -- Part VI. Helper lemmas for nonDecreasing
   --------------------------------------------------------------------------------------------
   nonDecreasingMerge <-
       inductWith cvc5 "nonDecreasingMerge"
@@ -604,12 +508,12 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
                             =: qed)
 
   --------------------------------------------------------------------------------------------
-  -- Part VI. Prove that the output of quick sort is non-decreasing
+  -- Part VII. Prove that the output of quick sort is non-decreasing
   --------------------------------------------------------------------------------------------
   sortIsNonDecreasing <-
      sInductWith cvc5 "sortIsNonDecreasing"
              (\(Forall @"xs" xs) -> nonDecreasing (quickSort xs))
-             (length @Integer) $
+             (length @a) $
              \ih xs ->
                 [] |- nonDecreasing (quickSort xs)
                    =: split xs trivial
@@ -647,10 +551,10 @@ correctness = runKDWith z3{kdOptions = (kdOptions z3) {ribbonLength = 60}} $ do
                                    =: qed)
 
   --------------------------------------------------------------------------------------------
-  -- Part VII. Putting it together
+  -- Part VIII. Putting it together
   --------------------------------------------------------------------------------------------
 
-  qs <- lemma "quickSortIsCorrect"
+  qs <- lemma (atProxy p "quickSortIsCorrect")
            (\(Forall @"xs" xs) -> let out = quickSort xs in isPermutation xs out .&& nonDecreasing out)
            [sortIsPermutation, sortIsNonDecreasing]