+++ CHANGES.md
@@ -10,15 +10,16 @@
       - Add a knuckle-dragger proof that tail-recursive definition of fibonacci is
         equivalent to its textbook definition. See Documentation/SBV/Examples/KnuckleDragger/Fibonacci.hs.
       - Renamed kdShowDepsHTML to showProofTreeHTML. (Along with showProofTree which renders in ASCII.)
+      - Add Documentation.SBV.Examples.KnuckleDragger.Reverse, proving a curious definition of reverse that
+        uses no auxiliary definitions is correct.
+      - Generalize various KD list/sort proofs.
 
   * Remove Data.SBV.Tools.NaturalInduction. The functionality provided by this tool is much better
     addressed by KnuckleDragger's induction methods. If you were using this functionality and have problems
     porting to knuckle-dragger, please get in touch!
 
-   * Add function 'last' to Data.SBV.List.
+  * Add functions 'last' and 'replicate' to Data.SBV.List.
 
-   * Add Documentation.SBV.Examples.KnuckleDragger.Reverse, proving a curious definition of reverse that
-     uses no auxiliary definitions is correct.
 
 ### Version 11.7, 2025-05-16